feat(hooks): move to package
