--- packages/hooks/src/index.ts
@@ -1,1 +1,0 @@
-export const name = 'hooks';
