+++ packages/hooks/src/use-is-mounted.ts
